moved auxiliar_functions, auxiliar_functions_gsl and kepler into dynamical_system
