--- include/auxiliar_functions_gsl.h
@@ -1,33 +1,0 @@
-#ifndef AUX_GSL_H
-#define AUX_GSL_H
-
-#include <math.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <gsl/gsl_odeiv2.h>
-
-#include "dynamical_system.h"
-
-/**
- * this library deals with functions that uses GSL
- * defined variables as arguments
-**/
-
-// sets driver for chosen error and control
-int set_driver(gsl_odeiv2_driver **d, 
-               gsl_odeiv2_system *sys, 
-			   const gsl_odeiv2_step_type *T, 
-			   double h, double h_max, double h_min,
-			   double error_abs, double error_rel,
-			   char *control);
-
-// sets the chosen integrator
-int set_integrator(const gsl_odeiv2_step_type **T, 
-                   char *integrator);
-
-// sets the chosen system
-int set_system(gsl_odeiv2_system *sys, 
-               dynsys system);
-
-#endif
